Simplifying isSafeSpot method a bit
Simplified the isSafeSpot method a bit. Should be a little less
restrictive, and actually more forward compatible. It now just checks
for a spot that is either empty or transparent, and makes sure the
platform isn't bad.

It has the same simplification that my last pull had, but I left in the
blockIDs this time.